+++ atom-core/src/main/java/org/mmfmilku/atom/Atom.java
@@ -1,0 +1,28 @@
+package org.mmfmilku.atom;
+
+
+import org.mmfmilku.atom.param.Param;
+import org.mmfmilku.atom.util.AssertUtils;
+
+import java.util.function.Function;
+
+@FunctionalInterface
+public interface Atom<T extends Param> extends Function<T, Boolean> {
+
+    Boolean execute(T param);
+
+    @Override
+    default Boolean apply(T t) {
+        return execute(t);
+    }
+
+    default Atom<T> before(Atom<? super T> before) {
+        AssertUtils.notNull(before);
+        return t -> before.execute(t) && execute(t);
+    }
+
+    default Atom<T> after(Atom<? super T> after) {
+        AssertUtils.notNull(after);
+        return t -> execute(t) && after.execute(t);
+    }
+}
