+++ atom-core/src/main/java/org/mmfmilku/atom/Process.java
@@ -1,0 +1,8 @@
+package org.mmfmilku.atom;
+
+@FunctionalInterface
+public interface Process {
+
+    void process();
+
+}
