+++ atom-core/src/main/java/org/mmfmilku/atom/param/Param.java
@@ -1,0 +1,4 @@
+package org.mmfmilku.atom.param;
+
+public interface Param {
+}
