+++ atom-core/src/main/java/org/mmfmilku/atom/param/operation/Copier.java
@@ -1,0 +1,7 @@
+package org.mmfmilku.atom.param.operation;
+
+public interface Copier<T, K> {
+
+    void copy(T param, K source, K target);
+
+}
