+++ atom-core/src/main/java/org/mmfmilku/atom/param/operation/Getter.java
@@ -1,0 +1,10 @@
+package org.mmfmilku.atom.param.operation;
+
+
+import org.mmfmilku.atom.param.Param;
+
+public interface Getter<T extends Param, K, V> {
+
+    V get(T param, K k);
+
+}
