+++ atom-core/src/main/java/org/mmfmilku/atom/param/operation/Setter.java
@@ -1,0 +1,9 @@
+package org.mmfmilku.atom.param.operation;
+
+import org.mmfmilku.atom.param.Param;
+
+public interface Setter<T extends Param, K, V> {
+
+    V set(T param, K k, V v);
+
+}
