+++ atom-core/src/main/java/org/mmfmilku/atom/util/AssertUtils.java
@@ -1,0 +1,10 @@
+package org.mmfmilku.atom.util;
+
+public class AssertUtils {
+
+    public static void notNull(Object o) {
+        if (o == null)
+            throw new NullPointerException();
+    }
+
+}
