+++ atom-core/src/main/java/org/mmfmilku/atom/util/AtomConst.java
@@ -1,0 +1,9 @@
+package org.mmfmilku.atom.util;
+
+import java.util.function.BiConsumer;
+
+public class AtomConst {
+
+    public static final BiConsumer NO_OPERATE = LambdaUtil.biNop();
+
+}
