+++ atom-parse/pom.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <artifactId>atom</artifactId>
+        <groupId>org.mmfmilku</groupId>
+        <version>0.0.1-SNAPSHOT</version>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>atom-parse</artifactId>
+
+    <dependencies>
+        <dependency>
+            <groupId>org.mmfmilku</groupId>
+            <artifactId>atom-core</artifactId>
+            <version>${atom-core-version}</version>
+        </dependency>
+    </dependencies>
+</project>
