+++ atom-parse/src/main/java/org/mmfmilku/atom/parser/Parser.java
@@ -1,0 +1,7 @@
+package org.mmfmilku.atom.parser;
+
+public interface Parser<T, R> {
+
+    R parse(T t);
+
+}
