--- src/main/java/atom/Process.java
@@ -1,8 +1,0 @@
-package atom;
-
-@FunctionalInterface
-public interface Process {
-
-    void process();
-
-}
