--- src/main/java/param/Param.java
@@ -1,4 +1,0 @@
-package param;
-
-public interface Param {
-}
