--- src/main/java/param/operation/Copier.java
@@ -1,7 +1,0 @@
-package param.operation;
-
-public interface Copier<T, K> {
-
-    void copy(T param, K source, K target);
-
-}
