--- src/main/java/param/operation/Getter.java
@@ -1,10 +1,0 @@
-package param.operation;
-
-
-import param.Param;
-
-public interface Getter<T extends Param, K, V> {
-
-    V get(T param, K k);
-
-}
