--- src/main/java/param/operation/Setter.java
@@ -1,9 +1,0 @@
-package param.operation;
-
-import param.Param;
-
-public interface Setter<T extends Param, K, V> {
-
-    V set(T param, K k, V v);
-
-}
