--- src/main/java/parser/Parser.java
@@ -1,7 +1,0 @@
-package parser;
-
-public interface Parser<T, R> {
-
-    R parse(T t);
-
-}
