--- src/main/java/util/AssertUtils.java
@@ -1,10 +1,0 @@
-package util;
-
-public class AssertUtils {
-
-    public static void notNull(Object o) {
-        if (o == null)
-            throw new NullPointerException();
-    }
-
-}
