--- src/main/java/util/AtomConst.java
@@ -1,9 +1,0 @@
-package util;
-
-import java.util.function.BiConsumer;
-
-public class AtomConst {
-
-    public static final BiConsumer NO_OPERATE = LambdaUtil.biNop();
-
-}
